# vk_test_friends/doc/openapi-manual.yaml
...
openapi: 3.0.0
info:
  title: VK Friends
  description: API для сервиса друзей
  version: 1.0.0
servers:
- url: http://localhost:8000
  description: localhost
...
paths:
...
  /api/auth/sing_up/verify_account/:
    post:
      tags:
      - auth
      description: Подтверждение регистрации
      operationId: verify_account
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyAccount'
        required: true
      responses:
        "200":
          description: Успешное подтверждение
          content:
            application/json:
              schema:
                type: object
              examples:
                УспешноеПодтверждение:
                  summary: Успешное подтверждение
                  value:
                    detail: Аккаунт успешно подтвержден
        "400":
          description: Ошибка подтверждения
          content:
            application/json:
              schema:
                type: object
              examples:
                ОшибкаПодтверждения:
                  summary: Успешный сброс пароля
                  value:
                    detail: Возникла ошибка при подтверждении сброса пароля. Проверьте корректность ссылки
...
components:
  schemas:
...
    VerifyAccount:
      required:
      - encoded_id
      - token
      type: object
      properties:
        token:
          type: string
          writeOnly: true
          example: bnpa4a-4255053c3097b23decf7849a7898f72c
        encoded_id:
          type: string
          writeOnly: true
          example: MjUy
      description: Схема для подтверждения регистрации
...
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT